+++ lib/main.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 
 import 'package:hydrus_elo/pages/authorization.dart';
+import 'package:hydrus_elo/pages/processing.dart';
 
 void main() {
   runApp(MaterialApp(initialRoute: '/authorization', routes: {
     '/authorization': (context) => Authorization(),
-    // '/home': (context) => Home(),
+    '/processing': (context) => Processing(),
     // '/location': (context) => ChooseLocation(),
   }));
 }